requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
python-dotenv==1.0.1
youtube-transcript-api==0.6.1
httpx==0.27.0
python-multipart==0.0.9
groq==0.4.1
requests==2.31.0
pytube==15.0.0
beautifulsoup4==4.12.3
lxml==4.9.3
SQLAlchemy==2.0.26
aiosqlite==0.19.0 